requirements.txt
requests
dnspython
python-whois
python-nmap
Jinja2
